requirements.txt
numpy==1.24.4
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
nfl-data-py
requests==2.31.0
pandas==2.0.3